extract ligoc from metrics repository into project repository, changes in usecase, injector and test
