+++ metrics/web/lib/common/data/repositories/firestore_project_repository.dart
@@ -1,0 +1,16 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:metrics/common/domain/repositories/project_repository.dart';
+import 'package:metrics_core/metrics_core.dart';
+
+/// An implementation [ProjectRepository] for [Firestore].
+class FirestoreProjectRepository implements ProjectRepository {
+  final Firestore _firestore = Firestore.instance;
+
+  @override
+  Stream<List<Project>> projectsStream() {
+    return _firestore.collection('projects').orderBy('name').snapshots().map(
+        (snapshot) => snapshot.documents
+            .map((doc) => ProjectData.fromJson(doc.data, doc.documentID))
+            .toList());
+  }
+}
