+++ metrics/web/lib/common/domain/repositories/project_repository.dart
@@ -1,0 +1,9 @@
+import 'package:metrics_core/metrics_core.dart';
+
+/// Base class for projects repositories.
+///
+/// Provides an ability to get the projects data.
+abstract class ProjectRepository {
+  /// Provides the stream of [Project]s.
+  Stream<List<Project>> projectsStream();
+}
