+++ metrics/web/lib/common/domain/usecases/receive_project_updates.dart
@@ -1,16 +1,16 @@
 import 'dart:async';
 
+import 'package:metrics/common/domain/repositories/project_repository.dart';
 import 'package:metrics/common/domain/usecases/usecase.dart';
-import 'package:metrics/dashboard/domain/repositories/metrics_repository.dart';
 import 'package:metrics_core/metrics_core.dart';
 
 /// Provides an the ability to receive [Project]s updates.
 class ReceiveProjectUpdates implements UseCase<Stream<List<Project>>, void> {
-  final MetricsRepository _repository;
+  final ProjectRepository _repository;
 
-  /// Creates the [ReceiveProjectUpdates] use case with the given [MetricsRepository].
+  /// Creates the [ReceiveProjectUpdates] use case with the given [ProjectRepository].
   ///
-  /// [MetricsRepository] must not be null.
+  /// [ProjectRepository] must not be null.
   const ReceiveProjectUpdates(this._repository) : assert(_repository != null);
 
   @override
