+++ metrics/web/lib/dashboard/domain/repositories/metrics_repository.dart
@@ -4,9 +4,6 @@ import 'package:metrics_core/metrics_core.dart';
 ///
 /// Provides an ability to get the metrics data.
 abstract class MetricsRepository {
-  /// Provides the stream of [Project]s.
-  Stream<List<Project>> projectsStream();
-
   /// Provides the stream of [Build]s of the project with [projectId]
   /// where elements are ordered by [Build.startedAt]
   /// and only last [limit] elements returned.
